requirements.txt
ultralytics
torch
torchvision
opencv-python-headless
numpy
flask
gunicorn
pillow
